## Summary
This PR optimizes the performance of the `Circuit.s_external()` method by introducing a auxiliary matrix `[T]` that balances the **assignment efficiency** of `F-contiguous` arrays and the **computational performance** of `C-contiguous` arrays.

## Background & Motivation
`Network.s` is a 3D `ndarray` (frequency, port1 and port2). When assigning `S-parameters`, `F-contiguous` layout (column-first) is faster because assignments are grouped by ports. However, operations such as `np.matmul()`, `np.linalg.solve()`, `np.linalg.inv()`, perform faster with `C-contiguous` arrays. This PR seeks to balance these trade-offs to improve both **assignment** and **computation** efficiency.

## Key Updates
1. Added `Circuit.X_F` and `Circuit.C_F`, which use the `F-contiguous` layout, in addition to the existing `Circuit.X` and `Circuit.C`. This significantly speed up assignment for the concatenated intersection matrix `[X]` and global scattering matrix `[C]`, resolving the numerical bottleneck mentioned in `Circuit.X`.
https://github.com/scikit-rf/scikit-rf/blob/6968484b38fc35fc1d7bff06aed343687ad537ba/skrf/circuit.py#L914-L930
   
2. Introduced an auxiliary matrix `[T]` to compute `Circuit.C @ Circuit.X`. Since `[T]` uses the `F-contiguous` versions of `Circuit.C_F` and `Circuit.X_F`, directly use of `@` would cause performance issue. To prevent this, the block diagonal structure of `Circuit.X_F` is exploited, using a for-loop to reduce the size of each matrix operation and increase assignment frequency, speeding up `[T]` calculations.

3. Improved type hinting and list comprehensions for cleaner, more concise code. While these changes don't significantly impact performance, they enhance code readability.

## Results
1. Using the case from [this PR](https://github.com/scikit-rf/scikit-rf/pull/1101), I analyzed the performance of `s_external()` from `v1.0.0` to this branck and generated the following chart:
![e5NJ9dJ6iP](https://github.com/user-attachments/assets/46e4ca92-2dec-4846-bd28-b1149722d345)
Benchmark code could find in this zip file.
[circuit_performance.zip](https://github.com/user-attachments/files/16868988/circuit_performance.zip)

2. Because the above case only includes **501 frequency points**, the performance improvement is not very pronounced. To provide a clearer comparison, I reran the test with my typical **2001 frequency points** (with `f_calls` = 100), and the performance comparison between `v1.3.0 (master)` and `this branch (CircuitPerformance)` is shown below:
    ```shell
    (skrf) ➜  scikit-rf git:(master) ✗ python circuit_performance.py
    Total time in Easy example (dim=10) => mean: 18.6318 ms, min: 14.0717 ms, std: 1.6515 ms
    Total time in Simple example (dim=18) => mean: 55.2932 ms, min: 41.8315 ms, std: 2.9023 ms
    Total time in HFSS example (dim=47) => mean: 363.5137 ms, min: 294.2424 ms, std: 27.1044 ms
    Total time in Complex example (dim=80) => mean: 997.8905 ms, min: 827.3208 ms, std: 79.8324 ms
    (skrf) ➜  scikit-rf git:(CircuitPerformance) ✗ python circuit_performance.py
    Total time in Easy example (dim=10) => mean: 18.9493 ms, min: 13.7978 ms, std: 2.0583 ms
    Total time in Simple example (dim=18) => mean: 57.4820 ms, min: 44.1387 ms, std: 5.8504 ms
    Total time in HFSS example (dim=47) => mean: 329.7286 ms, min: 255.8784 ms, std: 36.2242 ms
    Total time in Complex example (dim=80) => mean: 801.0437 ms, min: 623.7311 ms, std: 128.7105 ms
    ```