@Asachoo 2024-09-05
By optimizing the T property with `empty_like()` initialization and leveraging `np.einsum('...ii->...i', t)[:]` for trace extraction, we've achieved a further performance improvements.
In larger circuits scale, this enhancement has slashed the `s_external()` execution time by nearly 50%. (Due to time constraints, only `MKL` environment has been tested)
![Jsk0T9RCXM](https://github.com/user-attachments/assets/fe764645-ec50-4828-8e59-6c86c33d36e3)
![fM3yOkpZaV](https://github.com/user-attachments/assets/ff2e6917-f72b-40f6-af74-f9f4430fafce)